pyproject.toml
[tool.poetry]
name = "carsound_web"
version = "0.1.0"
description = ""
authors = ["Sangsoo Jin <constantlyjin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
django-environ = "^0.4.5"
django-cors-headers = "^3.6.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
django-maintenance-mode = "^0.15.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = "^5.3.1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
